+++ apps/plugin.h
@@ -192,7 +192,7 @@ struct plugin_api {
     void (*yield)(void);
     long* current_tick;
     int (*default_event_handler)(int event);
-    int (*default_event_handler_ex)(int event, void (*callback)(void *), void *parameter);
+    int (*default_event_handler_ex)(long event, void (*callback)(void *), void *parameter);
     int (*create_thread)(void* function, void* stack, int stack_size, const char *name);
     void (*remove_thread)(int threadnum);
     void (*reset_poweroff_timer)(void);
